requirements.txt
altair==5.4.1
amqp==5.2.0
asgiref==3.8.1
asttokens==2.4.1
attrs==24.2.0
beautifulsoup4==4.12.3
billiard==4.2.0
blinker==1.8.2
cachetools==5.5.0
celery==5.3.6
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
comm==0.2.2
contourpy==1.2.1
cowsay==6.1
crispy-bootstrap5==2024.10
cycler==0.12.1
debugpy==1.8.2
decorator==5.1.1
Django==5.0.6
django-bootstrap5==24.3
django-crispy-forms==2.3
django-debug-toolbar==4.4.6
django-extensions==3.2.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
et-xmlfile==1.1.0
executing==2.0.1
fonttools==4.53.0
gitdb==4.0.11
GitPython==3.1.43
idna==3.6
ipykernel==6.29.4
ipython==8.25.0
jedi==0.19.1
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.2
jupyter_core==5.7.2
kiwisolver==1.4.5
kombu==5.3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.0
matplotlib-inline==0.1.7
mdurl==0.1.2
narwhals==1.8.4
nest-asyncio==1.6.0
numpy==2.0.0
openpyxl==3.1.5
packaging==24.1
pandas==2.2.2
parso==0.8.4
pillow==10.3.0
platformdirs==4.2.2
prompt-toolkit==3.0.43
protobuf==5.28.2
psutil==6.0.0
psycopg2-binary==2.9.10
pure-eval==0.2.2
pyarrow==17.0.0
pydeck==0.9.1
Pygments==2.18.0
PyJWT==2.10.1
pyparsing==3.1.2
python-dateutil==2.8.2
python-decouple==3.8
pytz==2024.1
pywin32==306
pyzmq==26.0.3
referencing==0.35.1
requests==2.31.0
rich==13.8.1
rpds-py==0.20.0
six==1.16.0
smmap==5.0.1
soupsieve==2.5
sqlparse==0.5.0
stack-data==0.6.3
streamlit==1.38.0
tenacity==8.5.0
toml==0.10.2
tornado==6.4.1
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.1
vine==5.1.0
watchdog==4.0.2
wcwidth==0.2.13
